style(MemberInfoServiceTest) : BaseTest 상속함에 따른 @SpringBoot 애너테이션 삭제 및 의존 객체 나열 스타일 변경
